requirements.txt
RGBMatrixEmulator
Pillow
numpy
requests
spotipy
rgbmatrix
flask
pause
